Refactor: 폴더 구조 수정 #SCRUM-53
